Logic updates
If we are building a debug build, we're always working off of the dev
branch. If we are building a release build, we're always working off
of the release branch.

Merge dev into release when bumping the version.

Make updating the default; introduce -l flag for developers to avoid
losing local changes.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>